refactor: move mod panel logic into infractions/ dir
