+++ src/utils/moderation/modPanel/handlers/RemoveReactionsHandler.ts
@@ -1,7 +1,7 @@
 import { emojis } from '#utils/Constants.js';
 import type { ReactionArray } from '#types/Utils.d.ts';
 import { getOriginalMessage } from '#main/utils/network/messageUtils.js';
-import { ModAction, replyWithUnknownMessage } from '#utils/moderation/modActions/utils.js';
+import { ModAction, replyWithUnknownMessage } from '#main/utils/moderation/modPanel/utils.js';
 import { updateReactions } from '#utils/reaction/actions.js';
 import sortReactions from '#utils/reaction/sortReactions.js';
 import { ButtonInteraction, Snowflake } from 'discord.js';
